requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.1.2
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
comtypes==1.4.11
curl_cffi==0.14.0
db-to-sqlite==1.5
deprecation==2.1.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-yasg==1.21.10
email_validator==2.2.0
fake-useragent==2.2.0
fastapi==0.104.1
filelock==3.20.1
frozendict==2.4.7
frozenlist==1.7.0
fsspec==2025.12.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
gotrue==2.9.1
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.25.2
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.3.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.5.1
multitasking==0.0.12
mysql==0.0.3
mysql-connector-python==9.3.0
mysql-to-sqlite3==2.4.1
mysqlclient==2.2.7
networkx==3.6.1
numpy==2.2.5
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
peewee==3.18.3
pillow==11.2.1
platformdirs==4.5.1
pluggy==1.6.0
postgrest==0.15.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyJWT==2.10.1
pymongo==4.6.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.4
pytimeparse2==1.7.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
realtime==2.5.2
regex==2025.11.3
requests==2.31.0
rsa==4.9.1
safetensors==0.7.0
selenium==4.15.2
setuptools==80.3.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.10.4
SQLAlchemy==2.0.40
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.3
starlette==0.27.0
storage3==0.7.7
StrEnum==0.4.15
supabase==2.8.0
supafunc==0.3.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.23.2
watchfiles==1.1.1
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
yfinance==0.2.66
